Extend types from grafana/ui table side
