+++ packages/grafana-ui/src/components/Table/types.ts
@@ -4,15 +4,11 @@ import { CellProps, Column, Row, TableState, UseExpandedRowProps } from 'react-t
 
 import { DataFrame, Field, KeyValue, SelectableValue, TimeRange } from '@grafana/data';
 import { TableCellHeight } from '@grafana/schema';
+import * as schema from '@grafana/schema';
 
 import { TableStyles } from './styles';
 
-export {
-  type TableFieldOptions,
-  TableCellDisplayMode,
-  type FieldTextAlignment,
-  TableCellBackgroundDisplayMode,
-} from '@grafana/schema';
+export { type FieldTextAlignment, TableCellBackgroundDisplayMode } from '@grafana/schema';
 
 export interface TableRow {
   [x: string]: any;
@@ -91,3 +87,20 @@ export interface Props {
   /** @alpha Used by SparklineCell when provided */
   timeRange?: TimeRange;
 }
+
+export interface CustomCellRendererProps {
+  field: Field;
+  index: number;
+  frame: DataFrame;
+  value: any;
+}
+
+export interface TableCustomCellOptions {
+  cellComponent: FC<CustomCellRendererProps>;
+  type: schema.TableCellDisplayMode.Custom;
+}
+
+export type TableCellOptions = schema.TableCellOptions | TableCustomCellOptions;
+export type TableFieldOptions = Omit<schema.TableFieldOptions, 'cellOptions'> & {
+  cellOptions: TableCellOptions;
+};
